IGNITE-28946 Test fix + code review changes
