+++ modules/clients/src/test/java/org/apache/ignite/internal/processors/rest/JettyRestProcessorAbstractSelfTest.java
@@ -77,13 +77,10 @@ import org.apache.ignite.lang.IgniteBiPredicate;
 import org.apache.ignite.lang.IgnitePredicate;
 import org.apache.ignite.lang.IgniteUuid;
 import org.apache.ignite.testframework.GridTestUtils;
-import org.apache.ignite.testframework.junits.WithSystemProperty;
 import org.junit.Test;
 import org.junit.runner.RunWith;
 import org.junit.runners.Parameterized;
 
-import static org.apache.ignite.IgniteSystemProperties.IGNITE_ENABLE_OBJECT_INPUT_FILTER_AUTOCONFIGURATION;
-import static org.apache.ignite.IgniteSystemProperties.IGNITE_MARSHALLER_BLACKLIST;
 import static org.apache.ignite.IgniteSystemProperties.IGNITE_USE_BINARY_ARRAYS;
 import static org.apache.ignite.cache.CacheMode.PARTITIONED;
 import static org.apache.ignite.cache.CacheMode.REPLICATED;
@@ -110,7 +107,6 @@ import static org.apache.ignite.internal.processors.rest.GridRestResponse.STATUS
  */
 @SuppressWarnings("unchecked")
 @RunWith(Parameterized.class)
-@WithSystemProperty(key = IGNITE_ENABLE_OBJECT_INPUT_FILTER_AUTOCONFIGURATION, value = "false")
 public abstract class JettyRestProcessorAbstractSelfTest extends JettyRestProcessorCommonSelfTest {
     /** */
     private static boolean memoryMetricsEnabled;
@@ -133,8 +129,6 @@ public abstract class JettyRestProcessorAbstractSelfTest extends JettyRestProces
 
     /** {@inheritDoc} */
     @Override protected void beforeTestsStarted() throws Exception {
-        String path = U.resolveIgnitePath("modules/core/src/test/config/class_list_exploit_included.txt").getPath();
-        System.setProperty(IGNITE_MARSHALLER_BLACKLIST, path);
         System.setProperty(IGNITE_USE_BINARY_ARRAYS, Boolean.toString(useBinaryArrays));
 
         thirdPartyStore = new ConcurrentHashMap<>();
@@ -146,7 +140,6 @@ public abstract class JettyRestProcessorAbstractSelfTest extends JettyRestProces
 
     /** {@inheritDoc} */
     @Override protected void afterTestsStopped() throws Exception {
-        System.clearProperty(IGNITE_MARSHALLER_BLACKLIST);
         System.clearProperty(IGNITE_USE_BINARY_ARRAYS);
 
         super.afterTestsStopped();
@@ -181,7 +174,7 @@ public abstract class JettyRestProcessorAbstractSelfTest extends JettyRestProces
      * @param content Content to check.
      * @param err Error message.
      */
-    protected void assertResponseContainsError(String content, String err) throws IOException {
+    static void assertResponseContainsError(String content, String err) throws IOException {
         assertFalse(F.isEmpty(content));
         assertNotNull(err);
 
@@ -701,23 +694,6 @@ public abstract class JettyRestProcessorAbstractSelfTest extends JettyRestProces
         );
 
         assertResponseContainsError(ret, "Failed to convert value to specified type");
-
-        // Check forbidden type.
-        ForbiddenType forbidden = new ForbiddenType(new Exploit[] {
-            new Exploit(1),
-            new Exploit(2)
-        });
-
-        json = JSON_MAPPER.writeValueAsString(forbidden);
-
-        ret = content(DEFAULT_CACHE_NAME, GridRestCommand.CACHE_PUT,
-            "keyType", "int",
-            "key", "5",
-            "valueType", ForbiddenType.class.getName(),
-            "val", json
-        );
-
-        assertResponseContainsError(ret, "Deserialization of class " + Exploit.class.getName() + " is disallowed.");
     }
 
     /**
@@ -3137,44 +3113,6 @@ public abstract class JettyRestProcessorAbstractSelfTest extends JettyRestProces
                 .a('}');
 
             return sb.toString();
-        }
-    }
-
-    /** */
-    private static class ForbiddenType {
-        /** Data. */
-        @JsonProperty
-        private Exploit[] data;
-
-        /** */
-        ForbiddenType() {
-            // No-op.
-        }
-
-        /**
-         * @param data Data.
-         */
-        ForbiddenType(Exploit[] data) {
-            this.data = data;
-        }
-    }
-
-    /** */
-    private static class Exploit {
-        /** Value. */
-        @JsonProperty
-        private int val = 10;
-
-        /**
-         * @param val Value
-         */
-        Exploit(int val) {
-            this.val = val;
-        }
-
-        /** */
-        Exploit() {
-            // No-op.
         }
     }
