+++ modules/core/src/test/config/class_list_exploit_included.txt
@@ -18,4 +18,4 @@
 org.apache.ignite.spi.discovery.tcp.DiscoveryUnmarshalVulnerabilityTest$Exploit
 org.apache.ignite.stream.socket.SocketStreamerUnmarshalVulnerabilityTest$Exploit
 org.apache.ignite.internal.processors.rest.TcpRestUnmarshalVulnerabilityTest$Exploit
-org.apache.ignite.internal.processors.rest.JettyRestProcessorAbstractSelfTest$Exploit
+org.apache.ignite.internal.processors.rest.JettyRestProcessorClassFilterTest$Exploit
